# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Rocket backend API — social memory journal",
        "title": "Rocket API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/friends/invite": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send an app invitation via email or phone",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "Invite friend",
                "parameters": [
                    {
                        "description": "Email or phone",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/friend.InviteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/v1/friends/pulse": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get friends' online status (active now / recent / offline)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "Friend pulse",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/friend.PulseResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/v1/friends/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Search for users by name or handle",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "Search friends",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Max results (1-100, default 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/friend.SearchList"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/v1/friends/suggestions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get suggested users to add as friends",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "Friend suggestions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Max results (1-100, default 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/friend.SuggestionList"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/v1/friends/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a friend or cancel a pending request",
                "tags": [
                    "friends"
                ],
                "summary": "Remove friend",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Target user ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/v1/friends/{id}/request": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send a friend request to another user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "Send friend request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Target user ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/friend.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
        "/api/v1/friends/{id}/respond": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Accept or decline a pending friend request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "Respond to friend request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Requester user ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Action (accept|decline)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/friend.RespondRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/friend.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "friend.InviteRequest": {
            "type": "object",
            "required": [
                "email_or_phone"
            ],
            "properties": {
                "email_or_phone": {
                    "type": "string"
                }
            }
        },
        "friend.PublicUser": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "friend.PulseItem": {
            "type": "object",
            "properties": {
                "last_active_at": {
                    "type": "string"
                },
                "presence": {
                    "type": "string"
                },
                "status_text": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/friend.PublicUser"
                }
            }
        },
        "friend.PulseResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/friend.PulseItem"
                    }
                },
                "online_count": {
                    "type": "integer"
                }
            }
        },
        "friend.RespondRequest": {
            "type": "object",
            "required": [
                "action"
            ],
            "properties": {
                "action": {
                    "description": "\"accept\" | \"decline\"",
                    "type": "string"
                }
            }
        },
        "friend.SearchList": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/friend.SearchResult"
                    }
                },
                "next_cursor": {
                    "type": "string"
                }
            }
        },
        "friend.SearchResult": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "friendship_status": {
                    "type": "string"
                },
                "handle": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "friend.StatusResponse": {
            "type": "object",
            "properties": {
                "friendship_status": {
                    "type": "string"
                }
            }
        },
        "friend.Suggestion": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "reason": {
                    "$ref": "#/definitions/friend.SuggestionReason"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "friend.SuggestionList": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/friend.Suggestion"
                    }
                }
            }
        },
        "friend.SuggestionReason": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "mutual friend count when Type == \"mutual\"",
                    "type": "integer"
                },
                "type": {
                    "description": "\"mutual\" | \"suggested\"",
                    "type": "string"
                }
            }
        },
...
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}